requirements.txt
fastapi
uvicorn
redis
python-dotenv
pydantic